+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 - Unreleased
     - K Means now uses mini batch GD instead of SGD
+    - K Means in now an Online learner
     - Added Adjusted Rand Index clustering metric
     - Added Seeder Interface
     - Added Random and Plus Plus seeders
     - Changed order of K Means hyperparameters
     - Moved Ranking interface to anomaly detector namespace
     - N-gram Tokenizer now outputs ranges of word tokens
+    - Changed default Fuzzy C Means hyper-parameters
 
 - 0.0.10-beta
     - Added LODA online anomaly detector